execute CheckMojo in forked build
This allows the maven-plugin to specify that it wants to execute during
the `process-classes` lifecycle phase.

If a user runs `mvn com.spotify:missinglink-maven-plugin:check` from the
command-line on a freshly cloned repository, then this ensures that
missinglink is running on code that is actually compiled.

If the user instead has bound `missinglink:check` to a certain phase of
the lifecycle by adding an `<executions>` section to their pom.xml, then
this causes Maven to invoke some of the phases a second time, for
instance:

```
user wanted to do `mvn test` and plugin is bound to process-classes:
- generate-sources
- process-sources
- generate-resources
- process-resources
- compile
- process-classes
  - missinglink forks a new build:
    - generate-sources
    - process-sources
    - generate-resources
    - process-resources
    - compile
    - process-classes
      - missinglink:check runs
- generate-test-sources, test-compile, etc.
```

So the downside of this forking is that some steps are duplicated,
adding a tiny bit more time to the overall build.